modify

+++ algebra/number/Combination.cpp
@@ -52,7 +52,7 @@ long long combination_cached(long long n,long long r,long long m){
   r=min(r,n-r);
   return cache_fact[n]*cache_inv_fact[r]%m*cache_inv_fact[n-r]%m;
 }
-long long combination_mod_2(ll n,ll r){
+long long combination_mod_2(long long n,long long r){
   return n==(r|n-r);
 }
 long long homogeneous(long long n,long long r,long long m){
